fix: NAV  抽取
